test(ralph-loop): split session completion detector coverage
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
